fix styles and test
