+++ src/components/sidebar-navigation/SidebarNavigationItemStyles.tsx
@@ -92,3 +92,13 @@ export const SidebarNavigationFooter = styled.div`
   ${sidebarNavigationContentStyles}
 `;
 SidebarNavigationFooter.displayName = 'SidebarNavigationFooter';
+
+export const UnstyledListItem = styled.li`
+  all: unset;
+`;
+UnstyledListItem.displayName = 'UnstyledListItem';
+
+export const UnstyledUList = styled.ul`
+  all: unset;
+`;
+UnstyledUList.displayName = 'UnstyledUList';
