+++ src/components/sidebar-navigation/__tests__/SidebarNavigation-test.tsx
@@ -1,0 +1,41 @@
+/*
+ * Echoes React
+ * Copyright (C) 2023-2025 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+
+import { screen } from '@testing-library/react';
+import { renderWithMemoryRouter } from '~common/helpers/test-utils';
+import { SidebarNavigation } from '../SidebarNavigation';
+
+it('should have no a11y issues', async () => {
+  const { container } = renderWithMemoryRouter(<SidebarNavigation isCollapsed />);
+
+  await expect(container).toHaveNoA11yViolations();
+});
+
+it.each([
+  ['collapsed', true, 'true'],
+  ['expanded', false, 'false'],
+])('should render correctly when %s', (_, isCollapsed, expected) => {
+  renderWithMemoryRouter(<SidebarNavigation isCollapsed={isCollapsed} />);
+
+  expect(screen.getByLabelText('Secondary navigation')).toHaveAttribute(
+    'data-sidebar-collapsed',
+    expected,
+  );
+});
